Gut parsers module
