+++ microhapulator/tests/test_loading_functions.py
@@ -48,64 +48,6 @@ def test_load_marker_frequencies_missing_or_bad_columns(tsv):
         microhapulator.load_marker_frequencies(data_file(tsv))
 
 
-@pytest.mark.parametrize(
-    "tsv,nrows,markerid,offset",
-    [
-        ("def/acb-dozen-offsets.tsv", 42, "mh09USC-9qB", 193),
-        ("def/yellow-offsets.tsv", 21, "mh01KK-205", 207),
-    ],
-)
-def test_load_marker_definitions(tsv, nrows, markerid, offset):
-    markers = microhapulator.load_marker_definitions(data_file(tsv))
-    assert list(markers.columns) == ["Marker", "Offset", "Chrom", "OffsetHg38"]
-    assert markers.shape[0] == nrows
-    assert markers.Marker.iloc[10] == markerid
-    assert markers.Offset.iloc[10] == offset
-
-
-def test_load_marker_definitions_missing_column():
-    message = r"column\(s\) missing from marker definition file: Marker"
-    with pytest.raises(ValueError, match=message):
-        microhapulator.load_marker_definitions(data_file("def/orange-offsets-missing.tsv"))
-
-
-def test_load_marker_definitions_minimal():
-    defn_file = data_file("def/default-panel-offsets-minimal.tsv")
-    markers = microhapulator.load_marker_definitions(defn_file)
-    assert list(markers.columns) == ["Marker", "Offset"]
-    assert markers.shape[0] == 90
-
-
-@pytest.mark.parametrize(
-    "fasta,nseq,firstseq",
-    [
-        ("refr/acb-dozen-refr.fasta", 12, "mh07USC-7qB"),
-        ("refr/russ4-refr.fasta.gz", 4, "mh01KK-001"),
-    ],
-)
-def test_load_marker_reference_sequences(fasta, nseq, firstseq):
-    seqdict = microhapulator.load_marker_reference_sequences(data_file(fasta))
-    assert len(seqdict) == nseq
-    observed = next(iter(seqdict.keys()))
-    print(firstseq, observed)
-    assert firstseq == observed
-
-
-def test_cross_check_equal():
-    s1 = set(["mh19KK-056", "mh07NH-12", "mh06USC-6qD"])
-    s2 = set(["mh19KK-056", "mh07NH-12", "mh06USC-6qD"])
-    microhapulator.cross_check_marker_ids(s1, s2, "marker definitions", "marker frequencies")
-
-
-def test_cross_check_unequal():
-    s1 = set(["mh19KK-056", "mh07NH-12", "mh06USC-6qD"])
-    s2 = set(["mh19KK-056", "mh07NH-12"])
-    with pytest.raises(ValueError, match=r"marker IDs unique to set1={mh06USC-6qD};"):
-        microhapulator.cross_check_marker_ids(s1, s2, "marker definitions", "marker frequencies")
-    with pytest.raises(ValueError, match=r"marker IDs unique to set2={mh06USC-6qD};"):
-        microhapulator.cross_check_marker_ids(s2, s1, "marker definitions", "marker frequencies")
-
-
 def test_load_marker_filters():
     markers = ("mh01XYZ-1", "mh02XYZ-2", "mh02XYZ-3")
     configfile = data_file("filters.csv")
